rfc1071: Add "Internet Checksum" algorithm.
I first tried to implement RFC-1071 by translating their C code, but it
appears the C code is badly mangled.  The numerical examples weren't
helpful: I think I tried about 4 different implementations of this
algorithm and none of them seemed to agree with the numbers in RFC-1071.

But they agree with each-other, and with
http://www.netfor2.com/checksum.html so stuff it, I'll leave it at the
last one I found.